+++ src/kth/inda13/commandWorld/data/Location.java
@@ -2,7 +2,7 @@ package kth.inda13.commandWorld.data;
 
 public class Location extends BidimensionalConstruct {
 
-	public Location(int x, int y) {
+	public Location(double x, double y) {
 		super(x, y);
 	}
