+++ src/kth/inda13/commandWorld/data/Size.java
@@ -2,7 +2,7 @@ package kth.inda13.commandWorld.data;
 
 public class Size extends BidimensionalConstruct {
 
-	public Size(int x, int y) {
+	public Size(double x, double y) {
 		super(x, y);
 	}
